Changed Rect width and height types to int as everything from window dimensions to simulation objects will be on an integer range or grid

+++ Source/ElecCirSim/Rect.hpp
@@ -2,11 +2,11 @@
 
 typedef struct Rect
 {
-	double width;
-	double height;
+	int width;
+	int height;
 
 	Rect():
-		width(0.0),
-		height(0.0)
+		width(0),
+		height(0)
 	{}
 }Rect;
